Rearranged files and folders
