+++ tests/test_account.py
@@ -17,10 +17,10 @@
 #
 # The Initial Developer of the Original Code is
 # Mozilla.
-# Portions created by the Initial Developer are Copyright (C) 2011
+# Portions created by the Initial Developer are Copyright (C) 2010
 # the Initial Developer. All Rights Reserved.
 #
-# Contributor(s):
+# Contributor(s): Alin Trif <alin.trif@softvision.ro>
 #
 # Alternatively, the contents of this file may be used under the terms of
 # either the GNU General Public License Version 2 or later (the "GPL"), or
@@ -36,29 +36,40 @@
 #
 # ***** END LICENSE BLOCK *****
 
-from mozillians_page import MozilliansStartPage
+from pages.mozillians_page import MozilliansBasePage
+from pages.mozillians_page import MozilliansStartPage
 from unittestzero import Assert
-import pytest
-xfail = pytest.mark.xfail
 
-class TestInvite:
 
-    def test_inviting_an_invalid_email_address(self, mozwebqa):
+class TestAccount:
+
+    def test_login_with_invalid_credentials(self, mozwebqa):
         home_page = MozilliansStartPage(mozwebqa)
         login_page = home_page.click_login_link()
-        login_page.log_in()
-        invite_page = home_page.click_invite_link()
-        invite_page.invite("invalidmail")
-        Assert.true(invite_page.is_invalid_mail_address_message_present)
+        login_page.log_in("username@invalid.tld", "invalidpass")
+        Assert.true(login_page.is_invalid_credentials_text_present)
 
-    def test_invite(self, mozwebqa):
+    def test_login_with_invalid_ldap_credentials(self, mozwebqa):
         home_page = MozilliansStartPage(mozwebqa)
         login_page = home_page.click_login_link()
+        login_page.log_in("username@mozilla.com", "invalidpass")
+        Assert.true(login_page.is_invalid_credentials_text_present)
+
+    def test_login_logout(self, mozwebqa):
+        home_page = MozilliansStartPage(mozwebqa)
+        login_page = home_page.click_login_link()
+        Assert.true(login_page.is_csrf_token_present)
         login_page.log_in()
-        invite_page = home_page.click_invite_link()
-        Assert.true(invite_page.is_csrf_token_present)
-        mail_address = "validuser@example.com"
-        invite_success_page = invite_page.invite(mail_address)
-        Assert.true(invite_success_page.is_mail_address_present(mail_address))
-        Assert.true(invite_success_page.is_success_message_present)
-        Assert.true(invite_success_page.is_invite_another_mozillian_link_present)
+        Assert.true(home_page.is_logout_link_present)
+        login_page.click_logout_link()
+        Assert.true(home_page.is_login_link_present)
+
+    def test_reset_password(self, mozwebqa):
+        home_page = MozilliansStartPage(mozwebqa)
+        login_page = home_page.click_login_link()
+        password_reset_page = login_page.click_forgot_password_link()
+        Assert.true(password_reset_page.is_csrf_token_present)
+        Assert.true(password_reset_page.is_reset_password_button_present)
+        Assert.true(password_reset_page.is_email_field_present)
+        password_reset_page.reset_password()
+        Assert.true(password_reset_page.is_password_reset_sent_text_present)
